requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async and concurrency
asyncio-throttle==1.0.2
aiofiles==23.2.1
aiolimiter==1.1.0

# Data processing
numpy==1.24.3
pandas==2.0.3
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
datasets==2.14.6

# Document processing
pypdf==3.17.0
python-docx==1.1.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10

# Code analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.3
tree-sitter-java==0.20.2
tree-sitter-cpp==0.20.3

# Vector databases
qdrant-client==1.6.9
weaviate-client==3.25.3

# Search and retrieval
rank-bm25==0.2.2
faiss-cpu==1.7.4

# LLM clients
openai==1.3.5
anthropic==0.7.7

# Caching and storage
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Security and validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-benchmark==4.0.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Image processing and computer vision
opencv-python==4.8.1.78
clip-by-openai==1.0

# Table processing
tabula-py==2.8.2
camelot-py[cv]==0.11.0

# Environment and configuration
python-dotenv==1.0.0
click==8.1.7

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
tqdm==4.66.1
rich==13.7.0
typer==0.9.0
